backend/requirements.txt
Django==5.0
djangorestframework
psycopg2-binary
django-filter
djangorestframework-simplejwt
python-dotenv
